// applications/uk-ewsgit-settings/shared/types/settingType.ts
/*
 * Copyright ©2024 Ewsgit<https://github.com/ewsgit> and YourDash<https://github.com/yourdash> contributors.
 * YourDash is licensed under the MIT License. (https://ewsgit.mit-license.org)
 */

enum SETTING_TYPE {
  STRING = "string",
  INT = "int",
  FLOAT = "float",
  DIRECTORY = "directory",
  FILE = "file",
  FILE_OF_TYPE = "fileOfType",
  COLOR = "color",
  DATE = "date",
  TIME = "time",
  BOOLEAN = "boolean",
  SELECTION = "selection",
  MULTIPLE_SELECTION = "multipleSelection",
}

export default SETTING_TYPE;